+++ iterableapi/src/main/java/com/iterable/iterableapi/IterableConfig.java
@@ -86,9 +86,6 @@ public class IterableConfig {
      * By default, the SDK will save in-apps to disk.
      */
     final boolean useInMemoryStorageForInApps;
-
-    final boolean encryptionEnforced;
-
     /**
      * Allows for fetching embedded messages.
      */
@@ -109,7 +106,6 @@ public class IterableConfig {
         allowedProtocols = builder.allowedProtocols;
         dataRegion = builder.dataRegion;
         useInMemoryStorageForInApps = builder.useInMemoryStorageForInApps;
-        encryptionEnforced = builder.encryptionEnforced;
         enableEmbeddedMessaging = builder.enableEmbeddedMessaging;
     }
 
@@ -128,7 +124,6 @@ public class IterableConfig {
         private String[] allowedProtocols = new String[0];
         private IterableDataRegion dataRegion = IterableDataRegion.US;
         private boolean useInMemoryStorageForInApps = false;
-        private boolean encryptionEnforced = false;
         private boolean enableEmbeddedMessaging = false;
 
         public Builder() {}
@@ -258,17 +253,6 @@ public class IterableConfig {
         @NonNull
         public Builder setAllowedProtocols(@NonNull String[] allowedProtocols) {
             this.allowedProtocols = allowedProtocols;
-            return this;
-        }
-
-        /**
-         * Set whether the SDK should enforce encryption. If set to `true`, the SDK will not use fallback mechanism
-         * of storing data in un-encrypted shared preferences if encrypted database is not available. Set this to `true`
-         * if PII confidentiality is a concern for your app.
-         * @param encryptionEnforced `true` will have the SDK enforce encryption.
-         */
-        public Builder setEncryptionEnforced(boolean encryptionEnforced) {
-            this.encryptionEnforced = encryptionEnforced;
             return this;
         }